An example of BlazeSymbolizer.
addr2ln_sym needs the path of the image of the running kernel and an
ELF file.  It will print an address's symbol name and line information
in the userspace or the kernel space.

However, we don't provide the loaded address for the ELF file here, so
all userspace addresses are supposed to start from the address 0x0.

Signed-off-by: Kui-Feng Lee <kuifeng@fb.com>