ElfCache maintains a cache of Elf64Parser and DwarfResolver.
ElfCache has a capacity limitation.  It keeps the number of entries
under the limitation.  It purges old entries in LRU fashion.

ElfResolver calls ElfCache to get a backend, an Elf64Parser or a
DwarfResolver, of an ELF file.  ElfCache will try to create a new one
if there is not one already.

Signed-off-by: Kui-Feng Lee <kuifeng@fb.com>