Add an example of the symbolization against a process.
For an existing process, users can find a list of files mapped in the
process in /proc/<pid>/maps.  Looking into it, you can find where a
shared object; ex. libc.so, is loaded.  By giving the pid and an
address in the process, this exaple will find the symbol name, the
filename of the source, and the line number of the address.

For example,

	addr2ln_pid 19027 0x7f0c41ade000

This line will do symbolization for the address 0x7f0c41ade000 in the
process 19027.

Signed-off-by: Kui-Feng Lee <kuifeng@fb.com>