Remove the order check of CU start addresses.
Issue #8 raises the issue that the debug build will crash for checking
the order of CU start addresses.  However, this check is not valid
anymore since we sort the list of CUs.  In the future, we will support
returning multiple symbols and line numbers.  So, this check is either
unnecessary or too strict.

Signed-off-by: Kui-Feng Lee <kuifeng@fb.com>