Add benchmark for DWARF symbol parsing logic
Performance assessment seems to currently be happening via conditionally
compiled explicitly added time measurements in the code, printed to
stdout (enabled with the dwarf_perf feature).
This change introduces a proper benchmark for the one function currently
equipped with such bespoke timing instrumentation, debug_info_parse_symbols.

We rely on Rust's built-in benchmarking facility for this purpose.
Currently that is a nightly feature (requiring a nightly toolchain), but
to the best of my knowledge the test feature has not seen major changes
in the interface and we only use it for benchmarking anyway.

Alternatives include the bencher and criterion crates, which both work
on stable. The downside of both is that they mostly cater to
integration-test style benchmarks, which can only access the crate's
public API surface. However, for our intents and purposes it makes sense
to benchmark internal functions as well, in more unit-test style
benchmarks (using the #[bench] attribute instead of #[test]).

The purpose of this change is not so much to introduce comprehensive
infrastructure for tracking performance changes over time, possibly from
a CI environment and with a fancy UI, but mostly to have a way to
quickly quantify impact of changes in a before-after fashion in order to
be able to provide more data driven arguments about certain features or
code paths.
That being said, we do run the benchmark in CI, but mostly to curb bit
rot over time. The CI environment is slower (virtualized and less beefy)
and generally less predictable (because it's likely shared with other
tenants and unlikely to be fully partitioned off) than a local run,
despite all statistical gimmicks that the benchmark framework may
employ.

Signed-off-by: Daniel Müller <deso@posteo.net>