Remove dont_skip_tests feature
We don't need a feature if we don't want to skip a test based on a
runtime condition. We can just make it an ignored one, in which case it
can be run with `cargo test -- --ignored`.
Hence, remove the feature and mark the one test using it as ignored
instead. Note that currently the README contains broken code snippets. I
fixed it up to at least allow running of ignored tests, but it really
should be updated to use the up-to-date API.

Signed-off-by: Daniel Müller <deso@posteo.net>