Fix code snippets in README
The code snippets in the README are outdated. The SymSrcCfg type
mentioned there, for example, seems to have been renamed in the library
to SymbolSrcCfg eons ago. Update the README accordingly. Also tag Rust
code snippets as Rust code and C snippets as C. That has both the effect
that it will ensure proper handling of doc tests (C snippets are
ignored) while providing correct syntax highlighting. Note that not all
snippets can (or are intended to) compile, but those that can are fully
fixed up now and checked as part of CI by tagging them with `no_run`
instead of `ignore`.

Closes: #32

Signed-off-by: Daniel Müller <deso@posteo.net>