Use PathBuf/Path instead of String/str for representing paths
This change switches the library over to using PathBuf/Path instead of
String/str for representing paths. Strings are overly strong in their
assertions, in that they enforce UTF-8 encoding, which is not
necessarily required for paths. It's also the correct thing to do, as it
allows us to use native path manipulations. The conversion is fairly
straight forward and mechanical. It's also pretty much fully compiler
checked (modulo, perhaps, unsafe portions using pointer dances).

Signed-off-by: Daniel Müller <deso@posteo.net>