Remove Box usage and some safety issues from parse_debug_line_cu
The usage of Box::from_raw in parse_debug_line_cu is unlikely to be
safe: it requires that the caller made sure that the allocation being
referenced has exactly the same memory layout as Box would. That's not
a given here, for at least three reasons:
1) the allocation is coming from Vec, which had been resized to
  `prologue_sz`. This resize affects the vector's size, but not
  necessarily its capacity. However, for the purposes of allocation, it
  is the capacity that matters.
2) there is no guarantee of alignment: Vec<u8> may be 1 byte aligned,
   but that violates natural alignment (4 byte) that DebugLinePrologueV2
   requires, which in turn is in violation of the memory layout contract
   that Box::from_raw abides by
3) Rust is free to reorder fields in types, unless told not to. That has
   not happened, and DebugLinePrologue* members may not at all match up
   to what is in memory

(the latter is not a safety issues but a correctness one)

This change fixes these issues. For 3), we tag the DebugLinePrologue
types as repr(C), which prevents field reordering. Because we now use
ptr::read_unaligned, we just end up with an object on the stack and
allocator considerations are irrelevant, addressing 1). Similarly,
because we use an unaligned read into a Rust managed object on the
stack, we circumvent any alignment issues.

Signed-off-by: Daniel Müller <deso@posteo.net>