RawBufReader handles buffer offsets.
We use offsets to maintain the states of parsers intensively.
However, it is error-prone and cumbersome.  RawBufReader will take a
slice of &[u8] and maintain an offset.  It provides another copy of
the functions (leb128, word, dword, string...) that parse the basic
types we use.  Since we use the existing basic type parser functions
intensively, we keep both copies and gradually phase out the old ones.

Signed-off-by: Kui-Feng Lee <kuifeng@fb.com>