Make unsafe blocks explicit
It turns out we have heaps of implicit unsafe operations in unsafe
functions. Currently these don't require explicit wrapping in unsafe {}.
That has arguably been identified as a wrong design decision of Rust,
but it may only be changed in a new edition (if at all), because it
would be a breaking change.
The unsafe_block_in_unsafe_fn lint [0] has been devised to fix this
behavior, but it has to explicitly be opted in (though it may become the
default eventually). Let's deny it, forcing us to make all unsafe blocks
explicit. If we want to make any claims about memory safety of our code
base (the desire of which basically is one of the reasons why it is
being written in Rust), we need to have confidence that our unsafe code
is sane. That requires awareness of them first, followed by some form of
minimization, and auditing.
A not insignificant number of those blocks can presumably be removed
with a bit of work, but due to the sheer amount this will have to be
done in a step-by-step manner over time. This change only wraps all
unsafe operations in unsafe blocks in a mechanical manner.

[0] https://github.com/rust-lang/rfcs/blob/master/text/2585-unsafe-block-in-unsafe-fn.md

Signed-off-by: Daniel Müller <deso@posteo.net>