Introduce a binary to test on
A bunch of our tests work on the very unit test binary they are running.
That's a questionable practice as very little about it is well defined.
Needless to say, the moment we want to change anything the house of
cards falls apart. In the specific case at hand, if we decide to add
code coverage collection to the code base using cargo-llvm-cov, several
tests blow up for various reasons, the DWARF format being used in the
test binary being a major one.
To get at least some predictability in the process, this change
introduces the first test binary that we intend to work on in some of
the tests. The binary is compiled from source and no longer checked in.
It can be used for tests that don't make assumption about specific
number of debug elements or similar (which can change dramatically
between toolchains and compiler versions).

Signed-off-by: Daniel Müller <deso@posteo.net>