Introduce and use Addr typedef
We mix u64 and usize usage for representing addresses. This change
introduces the Addr typedef that is to be used moving forward. We define
that to map to usize, which is a machine word, which, on all but exotic
systems, equals the address size (C additionally has uintptr_t for that
purpose, but Rust decided not to have that and maps everything to
usize).
I adjusted all usage that I could identify. Everything I missed can be
cleaned up over time.

Signed-off-by: Daniel Müller <deso@posteo.net>