Make cbindgen dependency optional
The cbindgen dependency pulls in a swath of transitive dependencies.
Currently, everybody compiling the crate pays the price of compiling all
of them, while the dependency is really only used when generating C
headers.
With this change we pull in the dependency only when generating C
headers, eliminating that work when it is unnecessary.

Signed-off-by: Daniel Müller <deso@posteo.net>