Use dedicated test binaries in more tests
Some of our tests are currently broken on release builds, because they
work on the unit test binary itself, which may or may not provide the
properties being required by the test.
To fix this issue, this change introduces another binary for those tests
to run on and switches them over. The binary does not carry any debug
information, making sure that only ELF information is being used, where
that is possible.

Fixes: #44
Signed-off-by: Daniel Müller <deso@posteo.net>