Introduce ReadRaw trait for easily reading data
The current practice of reading data and then adjusting some read offset
based on implicit information is error prone. It leads to unnecessary
bugs such as https://github.com/libbpf/blazesym/issues/55, where we read
four bytes and then advance the "read pointer" by two. The existing read
functions are also infallible, which necessitates length check on each
and every call site and, if forgotten or done wrong, can be the cause of
panics. https://github.com/libbpf/blazesym/issues/55 illustrates that as
well.
This change introduces a saner API for reading data: the ReadRaw trait.
Currently implemented for byte slices, it provides fallible methods for
reading various primitive types and advancing the "read pointer" (the
start of the slice) accordingly. At this point we implement existing
decode_* functions in terms of this functionality.

Signed-off-by: Daniel Müller <deso@posteo.net>