Allow for opting out of test file generation
With upcoming changes we will require llvm-gsymutil to be available in
order to generate GSYM test files. Given that it is not installed by
default everywhere, at this point it probably makes to allow users to
opt out of the generation of test files for cases where it is clear that
tests are not meant to be run.
To make that possible, this change introduces a new feature,
dont-generate-test-files, that allows for just that. We enable that
feature for all CI jobs that don't run tests.

Signed-off-by: Daniel Müller <deso@posteo.net>