Start using ReadRaw trait in extract_attr_value()
With this change we start using the ReadRaw trait directly, converting
users of the existing non-fallible decode_*() set of functions to using
functionality provided by the trait. Specifically, we convert the
extract_attr_value() function over.

In so doing we also address part of the issues mentioned in
https://github.com/libbpf/blazesym/issues/55, because we remove the
explicit requirement for ahead-of-read length checks.

> src/dwarf/debug_info.rs | 292 ++++++++++----------------------------
> 1 file changed, 73 insertions(+), 219 deletions(-)

Signed-off-by: Daniel Müller <deso@posteo.net>