Remove measure_attr_size() function
As the comment over measure_attr_size() indicates, the function
effectively duplicates the logic of extract_attr_value() but skips the
actual reading, to the degree that is possible (it's not possible for
cases where we read a string or a variable sized integer). In so doing
it is prone to divergence from the "authoritative" logic, as evidenced
by the handling of the `DW_FORM_block4` variant:

The measure function returns size + 4:
        DW_FORM_block4 => {
            let sz = decode_uword(data);
            Some(sz as usize + 4)
        }

But the extract function returns size + 2:
        DW_FORM_block4 => {
            if 2 <= data.len() {
                let bytes = decode_uword(data);
                let fullsize = bytes as usize + 2;
                if fullsize <= data.len() {
                    Some((AttrValue::Bytes(&data[2..fullsize]), fullsize))
        ...

The reading is not particularly expensive in the context of the overall
operation, as indicated by the small performance difference reported by
our benchmarks:

- without reading:
  > test dwarf::tests::debug_info_parse_single_threaded ... bench:   4,488,850 ns/iter (+/- 137,953)
- with reading:
  > test dwarf::tests::debug_info_parse_single_threaded ... bench:   4,671,883 ns/iter (+/- 139,215)

This change removes it, leaving only extract_attr_value() in place and
making callers use that function. We can revisit how to speed up these
code paths if the need arises, but there likely are better approach than
two duplicated implementations.

Signed-off-by: Daniel Müller <deso@posteo.net>