Stop returning number of bytes read from core extract_attr_value() logic
This change splits the extract_attr_value() function into
extract_attr_value_impl() and extract_attr_value(). The former contains
the existing "core" logic, but no longer reports the number of bytes
read.
Instead, this number is now calculated in a single location. In so doing
we fix https://github.com/libbpf/blazesym/issues/55, the remaining issue
of which was that it reported the wrong number of bytes read to callers.
With the current approach, which combines reading with bounds checking
and the advancement of the read pointer, such issues are generally
mitigated.

Fixes: #55
Signed-off-by: Daniel Müller <deso@posteo.net>