Memory map ELF file
This change marks the first step towards using a memory mapped file when
parsing ELF objects. Doing so is more performant, and requires less
complex code, and will be more flexible down the line (e.g., if we
decide to switch to gimli for DWARF support).
Start off by mapping the file and storing it in a member variable.

Signed-off-by: Daniel Müller <deso@posteo.net>