Enable test file generation when building tests
Cargo build scripts can't distinguish between building the "core"
library and building tests. For that reason, we decided to make test
file generation opt-out.
It turns out there actually is a way to approximate the desired behavior
of building test files when testing very closely and it is outlined in
the corresponding upstream feature request:
https://github.com/rust-lang/cargo/issues/2911
Basically, we declare blazesym as a dev-dependency of itself with the
feature in question enabled.
With this change we do just that for the newly introduced
generate-test-files feature. That means test file generation is now
opt-in for regular builds, unless tests or examples are also build, in
which case it will default to being enabled. For that reason, we keep
the dont-generate-test-files feature around to allow for opting out of
that (unlikely to be needed often, but for instance useful for linting
the entire crate including tests and examples in CI without needing
llvm-gsymutil, for example).

Signed-off-by: Daniel Müller <deso@posteo.net>