Unseal Pod trait
When we introduced the Pod trait, which identifies "plain old data"
types, the intention was for it to be used only for a limited set of
built-in integer types. However, it actually makes sense to allow other
types used in the crate, such as those for ELF and DWARF definitions, to
implement this trait as well, which will allow us to more easily read
values of those types in our parsing code.
To make that happen this change "unseals" the trait, making it available
for implementation throughout the crate.

Signed-off-by: Daniel Müller <deso@posteo.net>