Rework ElfParser::ensure_ehdr() to use memory mapped file
This change adjusts the ElfParser::ensure_ehdr() method to use the
memory mapped file instead of reading data using explicit IO. The method
is moved into the Cache type, which is where it belongs. Also, the
method returns the ELF header reference directly, to prevent more
spreading of yet another error-prone pattern where we first "ensure"
that an object is present and then unwrap it. Not all such unwraps are
guaranteed to be preceded by the ensure call, which can lead to panics.

Signed-off-by: Daniel Müller <deso@posteo.net>