Rework ElfParser::ensure_shdrs() to use memory mapped file
This change adjusts the ElfParser::ensure_shdrs() method to use the
memory mapped file instead of reading data using explicit IO. The method
is moved into the Cache type, which is where it belongs. In addition to
that, use proper checked array accesses *when the access happens* and
not at an arbitrary time beforehand where it may easily be forgotten.

Signed-off-by: Daniel Müller <deso@posteo.net>