Add support for reading references of PODs
With upcoming changes we will move the ELF parser to a more zero-copy
inspired operating model. As such, we would like to just reference parts
of a memory mapped ELF file instead of copying them around.
To enable this use case, this change extends the ReadRaw trait with a
new method, read_pod_ref(), that can be used to (hopefully?) safely read
such a reference.

Signed-off-by: Daniel Müller <deso@posteo.net>