Rework ElfParser::ensure_phdrs() to use memory mapped file
This change adjusts the ElfParser::ensure_phdrs() method to use the
memory mapped file instead of reading data using explicit IO. The method
is moved into the Cache type, which is where it belongs. Also remove
some unnecessary `unsafe` code, which was entirely unsound, because it
(entirely needlessly!) created a mutable reference from a shared
context.

Fixes: #63
Signed-off-by: Daniel Müller <deso@posteo.net>