Emit blazesym.h header file into include/ directory
This patch adjusts the location of the generated blazesym.h header file
to reside in the include/ directory as opposed to target/<build-type>.
With upcoming changes the file will be checked in and having it in
include/ seems more intuitive. Also, there is little point in having it
in a build type specific directory when it is the same across build
types.

Signed-off-by: Daniel Müller <deso@posteo.net>