Include blazesym-benchmarks into repository
I originally created blazesym-benchmarks [0] as a separate repository
because we wanted to include and benchmark on "real world" data, which
was rather big. But given that the data compresses rather well (700 MiB
image down to ~100 MiB file) and that we now use git-lfs to manage this
data outside of the repository proper and that checkout of it is
optional, it arguably makes more sense to include these benchmarks here
right away instead. Doing so has the advantage that benchmarks are
co-located with library code in a single repository.
In order to run this new set of benchmarks the generate-bench-files
feature must be present. The benchmarks take a while to run and are
generally run much less frequent than tests, so an opt-in scheme seems
much more apt than what we have for tests where we generate test files
by default in dev-builds. That's also done to not *require* users to
have git-lfs installed.
We use criterion as opposed to Rust's built-in benchmark here. The
former provide better configurability and more rigorous statistical
analysis. The downside is that integration is not quite as nice (with a
bunch of macro magic), but it is function over form in this instance.

[0] https://github.com/danielocfb/blazesym-benchmarks

Signed-off-by: Daniel Müller <deso@posteo.net>