Add end-to-end test for DWARF symbol lookup
A recent commit added an end-to-end test for symbolizing an address
using DWARF debug information. This change introduces another test for
the reverse operation: retrieving an address for a symbol.

Note that our GSYM resolver currently does not support such a lookup,
which is why we cannot add a test for this format.

Signed-off-by: Daniel Müller <deso@posteo.net>