Directly reference FileInfo data
Instead of all this dance we currently do in GsymContext::file_info(),
parse FileInfo objects directly and then just index into file_tab. We
should make sure that when parsing the header, the read pointer is
properly aligned, which is done in the reference implementation but was
missing so far.

Signed-off-by: Daniel Müller <deso@posteo.net>