Use standard binary search for address search
Brought up as part of a review a while back [0], this change switches
the find_address() function over to using Rust's standard
binary_search() (in the form of partition_point(), which is just a thin
wrapper around it) instead of rolling our own. The new utility function
is generally useful and can be used for replacing more of our
hand-rolled binary searches in the future.

[0] https://github.com/libbpf/blazesym/pull/54#discussion_r1116119283

Signed-off-by: Daniel Müller <deso@posteo.net>