Eliminate nix dependency
This change remove the nix dependency from the crate, now that we no
longer rely on any functionality it provides. Apparently the crate was
responsible for Debug implementations of the various ELF related data
types used in our Cache type, so we now have to provide our own
Debug implementation for it and can no longer rely on a derive.

Signed-off-by: Daniel Müller <deso@posteo.net>