Rework map file parsing
Our current approach to parsing the /proc/.../maps file is problematic
in two ways:
1) it relies on the regex dependency which is something we want to avoid
   in this part of the code base, as it in turn contains additional
   transitive dependencies that some of our users don't want; it's also
   entirely unnecessary to use a regular expression for this file: its
   format was deliberately chosen to *not* require a regular expression
   but can be parsed using simple string operations such as splits at
   white spaces
2) the regular expression used is unnecessarily restrictive in the
   strings it accepts; e.g., if a hex address is emitted with upper case
   letters or the mode string contains a character other than rwxp-, it
   will not accept the line. In fact, this component seems to also
   support 's' (shared) [0], which we would be unable to recognize and would
   fail to parse.

With this change we switch over to using simple string operations for
parsing the contents. We are more flexible in what we accept and so the
above bug for shared mappings is no longer present. This change also
fixes a bug where we were emitting a wrong path in the parsed entry for
`(deleted)` entries when invoked on the "self" process. On top of that,
error reporting is improved by getting rid of various potential panics.
Lastly, actual tests are added for this functionality.

[0]: https://www.kernel.org/doc/html/latest/filesystems/proc.html#id10

Signed-off-by: Daniel Müller <deso@posteo.net>