Introduce zip module
In the not-so-distant future we intend to support symbolization of
addresses in APKs, which are really only zip files with certain
conventions.
In order to support that, this change introduces a zip module that can
be used to work with such archives. We may adjust the interface some
more moving forward, but for now we essentially just translated the
corresponding C code. See commit 1eebcb60633f ("libbpf: Implement basic
zip archive parsing support") [0] for the corresponding libbpf change.

[0]: https://github.com/torvalds/linux/commit/1eebcb60633fd469ee27b0fbd7ee4f271feedeca

Signed-off-by: Daniel Müller <deso@posteo.net>