Make iteration over zip entries possible
The current API of the zip Archive type, coming from a direct
translation of the C API, does not lend itself to Rust usage patterns.
For example, it's not possible to iterate over the entries in such an
archive.
This change introduces an iterator type that allows for iteration over
said entries. We will require this functionality once we start support
symbolizing addresses pointing within an mmap'ed archive.

Signed-off-by: Daniel Müller <deso@posteo.net>