Eliminate unnecessary rebuilds in build script
Our various test and benchmark artifacts are created repeatedly on every
build. While not a noticeable slowdown at this point, it's unnecessary
work and, over time, the cause of wear for SSDs, for example.

The problem seems to be that the generated files' mtimes are after those
of Cargo's output file that they are compared against to determine
staleness. This behavior is counter-intuitive: it is the build script
that emits those outputs to stdout while also generating those files.
Hence, the build script's outputs can only be available completely after
said script has run (that is, after the files have been generated as
well). Alas, this behavior seems to be fully intentional from the Cargo
side [0].

With this change we work around this problem, by adjusting generated
files' time stamps to those of said output file after generation. It's
not perfect (the output file is more or less an implementation detail
and could potentially change at some point; also, because we rely on the
existence of this output file we need two builds before the mtimes are
correct), but should be good enough. We also add a CI job that makes
checks that we don't re-generate files for the default
generate-test-files feature.

[0] https://github.com/rust-lang/cargo/blob/4a60ff78a5f03e7c9ed2399754efb62c80a8ecc8/src/cargo/core/compiler/fingerprint/mod.rs?L249#L249-L256

Fixes: #91
Signed-off-by: Daniel Müller <deso@posteo.net>