Untangle cache holder logic
The "CacheHolder" argument to KernelResolver::new() serves no real
purpose. All we do in there is to query this "holder" for the ksym cache
to retrieve the corresponding resolver. If we need a resolver, let's
just pass the resolver in.
Furthermore, it makes no sense for KernelResolver to store the kallsyms
path when the underlying resolver already contains this path as well.

Signed-off-by: Daniel Müller <deso@posteo.net>