Introduce proper logging facility
Introduce a proper logging facility, instead of having println
statements scattered throughout the code. This way, clients are given
the ability to control how to emit log messages. The feature is optional
and off by default, except for dev builds, which enable it. Usage for
testing can be done as follows:
  $ RUST_LOG=trace cargo test -- --nocapture

Signed-off-by: Daniel Müller <deso@posteo.net>