Add support for symbolizing addresses in APKs
This change adds support for symbolizing addresses in APKs, which has
relevance particularly for usage on Android.
To make that happen we create ELF resolvers for the individual ELF files
in such an APK as part of the work of building the process virtual
address space map. This is a bit of a kludge currently, because ideally
we would not eagerly build resolvers for all entries but do so lazily
for the relevant address ranges. We also currently don't use the
ElfCache (or any kind of caching for that matter), because the entire
logic for creating resolvers needs to be reworked from an address range
based one to a path lookup that is prefixed with an address
normalization step.
This change fixes what appears to be a bug in the ELF resolver, where
the offsets are not handled correctly, currently.

Signed-off-by: Daniel Müller <deso@posteo.net>