Allow for early filtering of proc maps entries
This change refactors the parsing logic for proc maps slightly, allowing
for passing in of a filter function that allows for early discarding of
entries that we have no interest in dealing with. Because we move the
generalized existing filter functionality into the maps module, we make
it easier reusable for future changes (which will require it). On top of
that such early filtering shrinks the size of allocations used.

Signed-off-by: Daniel Müller <deso@posteo.net>