Introduce dedicated types for SymbolSrcCfg variants
The variants of the SymbolSrcCfg enum currently do not have dedicated
types -- they are just defined inline. That makes it potentially hard to
work with them. For example, we can't pass only the kernel variant to a
function.
With this change we introduce dedicated types representing these
variants. This way, they can be named and worked with in isolation.

Signed-off-by: Daniel Müller <deso@posteo.net>