Remove CacheHolder type
The CacheHolder type is an arguably unnecessary construct that just
complicates control flows: instead of creating and passing in the cache
type that is required we now have to create an object of this
encompassing type and work with it.
Let's be explicit here instead and deal with the respective caches
directly.

Signed-off-by: Daniel Müller <deso@posteo.net>