Expose data offset in zip entries
With upcoming changes we want to access ELF files contained in an APK
archive that has been memory mapped. To make that happen via a shared
Mmap object, we need to know the offset at which the file resides.
With this change we expose this data inside the Entry type.

Signed-off-by: Daniel Müller <deso@posteo.net>