Make zip::Archive::entries() infallible
The zip::Archive::entries() method currently can fail, because the
central directory offset could potentially be out of bounds. That makes
for a bit awkward usage. Let's make sure to validate the offset during
construction of the object to eliminate this error path. A zip without a
valid central directory is effectively useless.

Signed-off-by: Daniel Müller <deso@posteo.net>