Add logic for reading KASLR offset
In order to (eventually...) support normalization of kernel addresses,
we need to take into account whether the running kernel has address
space layout randomization enabled. If that is the case, we will need to
incorporate the randomization offset in the normalization process. This
change introduces the necessary logic for reading said offset, so that
it can be used down the line.
This change builds on all the infrastructure we added for making the ELF
parser optionally work with using regular I/O APIs instead of relying on
memory mapping.

Refs: #950

Signed-off-by: Daniel Müller <deso@posteo.net>