Add support for normalizing user space addresses
We want to support symbolization in scenarios where addresses are
captured on one system and then symbolized elsewhere (#61). In a
nutshell, we are going to perform address "normalization" on the "local"
system (the one that captured the addresses), send them to the remote,
and then symbolize these normalized addresses on the remote.
This change introduces the logic for normalizing addresses. Basically,
we use the corresponding proc maps entry, parse the contents, identify
the ELF files to which those addresses map, and then perform
normalization based on the ELF information.
I introduced a new public module, normalize. We have to think some more
about how best to expose this functionality, but this is a reasonable
starting point.
The normalization does not yet use advanced caching of sorts: each ELF
entry handled is parsed as we found it. That is not particularly
effective but not wrong. We can optimize further subsequently without
adjusting the API.

Signed-off-by: Daniel Müller <deso@posteo.net>