Read and populate build ID fields
As part of the address normalization, users may needs the build ID of
the binary in which an address lies along with the binary's path. That's
useful in cases where a remote server is queried which indexes binaries
by build ID, such as with debuginfod.
With this change we make sure to attempt to read the build ID of the ELF
binaries that we found and populate the corresponding information
inside UserAddrMeta.

Signed-off-by: Daniel Müller <deso@posteo.net>