Add support for creating "views" on memory mappings
Capabilities for sharing memory mappings created at the library level
have been introduced earlier. That was necessary to support, for
example, parsing an ELF file contained in a (uncompressed memory mapped)
zip archive.
However, retrieving a "view" on some part inside said mapping was rather
ad-hoc and resulted in discussions [0].
With this change we introduce a proper API for restricting the view on a
memory mapping, while preserving the shareability of the underlying OS
memory. Basically, Mmap is now an Rc internally (future changes we
remove existing usages of Rc<Mmap>) and contains the constrain() method
that creates a new Mmap object with a smaller view.

[0] https://github.com/libbpf/blazesym/pull/110#discussion_r1163432667

Signed-off-by: Daniel Müller <deso@posteo.net>