Remove file offset lookup functionality
Blazesym currently "supports" looking up file offsets when looking up
addresses for a name. That's really only implemented for the ELF
resolver and nothing else. And that's arguably the only one for which it
has any meaning. That in turn suggests it's a concept not general enough
to expose in the resolver API to begin with. Remove it.

Signed-off-by: Daniel Müller <deso@posteo.net>