Fix end-to-end benchmark builds
The end-to-end benchmarks broke because of some API change a while back.
Fix them. Unfortunately, it's not easy to just build them in CI, because
at least currently, that is bound to the extraction of our large
vmlinux-5.17.12-100.fc34.x86_64 image, which we do not want to pull in
CI.

Signed-off-by: Daniel Müller <deso@posteo.net>