Introduce 'symbolize' feature
With this change we introduce the 'symbolize' feature, which guards all
the symbolization related functionality. It is enabled by default. When
disabled, mostly the normalization part remains, which can be useful in
remote symbolization scenarios, for capturing and normalizing addresses
on embedded devices, for example. When this feature is disabled, the
regex and lru dependencies are no longer necessary, meaning that all we
need is libc.

Signed-off-by: Daniel Müller <deso@posteo.net>