Remove file system based "is file" check from is_symbolization_relevant()
The proc maps logic for filtering of relevant entries currently makes a
system call to get file system information about whether the provided
path is a file or not. It's not entirely clear why we need that: if the
path starts with a slash (which we already check) it will presumably
always be a file, or at least it's not clear what else it could be. Omit
this addition check via system call.

Signed-off-by: Daniel Müller <deso@posteo.net>