Don't require generate-bench-files for all benches/
Currently all our end-to-end benchmarks in benches/ require the
generate-bench-files feature. In the future we will add more benchmarks
that will not require this feature.
To make that possible, rework everything to check for the presence of
the feature at run time. This has the added benefit that problems such
as the build failures fixed by commit fbecfdd39baa ("Fix end-to-end
benchmark builds") can no longer creep in, because we can now build the
benchmarks as part of CI.

Signed-off-by: Daniel Müller <deso@posteo.net>