Make symbolize call only take a single symbol source/configuration
There is not much of a point in passing a set of symbol
sources/configurations to a symbolize call. Rather, it only complicates
a bunch of things, from the API to the various internals. For one, now
we need to create this "resolver map" in order to find the "correct"
resolver to use. That in turn requires each resolver to know the range
it covers -- which it otherwise would not have to. It also means that
users are asked to provide a base address for sources such as gsym or
ELF.
With this change we only accept a single symbol source/configuration and
adjust all APIs accordingly.

Signed-off-by: Daniel Müller <deso@posteo.net>