Rework process symbolization logic
We want to transition away from the ResolverMap type and towards relying
on the existing normalization functionality as part of our symbolization
matters.
As a first step, this change implements process symbolization in terms
of said normalization functionality. Now that we pass in proper
normalized addresses to the ElfResolver, we can fix it's broken
find_symbols() logic.

Signed-off-by: Daniel Müller <deso@posteo.net>