Remove BlazeSymbolizer::find_address_regex() support
We are in the process of repackaging the "find address" functionality,
exposing it through a different API.
This change removes the regular expression lookup support associated
with it. We currently have no actual use cases for it, at best minimal
tests, and as such this is a bunch of code that we would have to convert
over and repackage for no really good reason. We can resurrect these
pieces in the future, should the need arise.

Signed-off-by: Daniel Müller <deso@posteo.net>